requirements.txt
RPi.GPIO==0.7.0
SimpleMFRC522==1.0.0